Final GUI Update

+++ src/Page_Home_Admin.java
@@ -1,5 +1,7 @@
 package src;
 
+import javax.swing.JOptionPane;
+
 import javafx.geometry.Pos;
 import javafx.scene.Scene;
 import javafx.scene.control.Button;
@@ -16,11 +18,12 @@ public class Page_Home_Admin extends BorderPane {
     TableView<obj_User> table;
 
     public Page_Home_Admin(Stage stage, String User, Scene scene) {
+        this.setId("HomePage");
         utils_newLabel lbl = new utils_newLabel();
         Label lusername = lbl.createNewLabel("Welcome. " + User + "!");
         lusername.setId("User");
-        HBox welcom = new HBox(30.0, lusername);
-        VBox Header = new VBox(20.0, lbl.createNewTitle("HomePage"), welcom);
+        HBox welcom = new HBox(10.0, lusername);
+        VBox Header = new VBox(10.0, lbl.createNewTitle("HomePage"), welcom);
         table = newtable.createUserTable();
         this.setCenter(table);
 
@@ -28,7 +31,7 @@ public class Page_Home_Admin extends BorderPane {
         Button deleteBtn = new Button("Delete");
         Button saveBtn = new Button("Save to File");
         Button logoutBtn = new Button("Logout");
-        VBox buttonBar = new VBox(10.0, createBtn, deleteBtn, saveBtn, logoutBtn);
+        HBox buttonBar = new HBox(10.0, createBtn, deleteBtn, saveBtn, logoutBtn);
         createBtn.setOnAction((evt -> {
             fm.registerNewUser();
             table.getItems().clear();
@@ -38,18 +41,32 @@ public class Page_Home_Admin extends BorderPane {
             fm.updateFile(table.getItems(), "User");
         }));
         logoutBtn.setOnAction((evt -> {
-            stage.setScene(scene);
-            stage.setTitle("ZooInsight - Login");
+            int logout = JOptionPane.showConfirmDialog(null, "Do you want to log out?", "Loging Out",
+                    JOptionPane.YES_NO_OPTION,
+                    JOptionPane.QUESTION_MESSAGE,
+                    null);
+            if (logout == 0) {
+                stage.setScene(scene);
+                stage.setTitle("ZooInsight - Login");
+            }
         }));
         deleteBtn.setOnAction((evt -> {
-            newtable.deleteEntry(table);
-            fm.updateFile(table.getItems(), "User");
+            int check = JOptionPane.showConfirmDialog(null, "Do you want to delete selected row?", "Delete",
+                    JOptionPane.YES_NO_OPTION,
+                    JOptionPane.QUESTION_MESSAGE,
+                    null);
+            if (check == 0) {
+                newtable.deleteEntry(table);
+                fm.updateFile(table.getItems(), "User");
+                JOptionPane.showMessageDialog(null, "User account deleted.", "Saved", JOptionPane.INFORMATION_MESSAGE,
+                        null);
+            }
         }));
 
-        this.setRight(buttonBar);
+        this.setBottom(buttonBar);
         this.setTop(Header);
         buttonBar.setId("Btnbar");
-        buttonBar.setAlignment(Pos.CENTER);
+        buttonBar.setAlignment(Pos.TOP_CENTER);
         welcom.setAlignment(Pos.TOP_CENTER);
         utils_background bckgrnd = new utils_background();
         bckgrnd.setBgrd("resources\\background.png");
